requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.0
cryptography==42.0.7
sentence-transformers==3.0.0
scikit-learn==1.4.2
apscheduler==3.10.4
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.7.1
